siamese_network/webapp/requirements.txt
flask==2.3.3
werkzeug==2.3.7
tensorflow>=2.12.0
numpy>=1.23.5
pillow>=9.5.0
scipy>=1.10.1
matplotlib>=3.7.2